+++ lib/src/controls/navigation/tab_view.dart
@@ -1,30 +1,67 @@
 import 'package:fluent_ui/fluent_ui.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/rendering.dart';
+import 'package:flutter/services.dart';
 
+/// The TabView control is a way to display a set of tabs
+/// and their respective content. TabViews are useful for
+/// displaying several pages (or documents) of content while
+/// giving a user the capability to rearrange, open, or close
+/// new tabs.
+///
+/// ![TabView Preview](https://docs.microsoft.com/en-us/windows/uwp/design/controls-and-patterns/images/tabview/tab-introduction.png)
+///
+/// See also:
+///   - [NavigationPanel]
 class TabView extends StatelessWidget {
+  /// Creates a tab view.
+  ///
+  /// [tabs] must have the same length as [bodies]
   const TabView({
     Key? key,
     required this.currentIndex,
     this.onChanged,
     required this.tabs,
     required this.bodies,
-    this.showNewButton = true,
     this.onNewPressed,
     this.addIconData = Icons.add,
+    this.shortcutsEnabled = true,
   })  : assert(tabs.length == bodies.length),
         super(key: key);
 
+  /// The index of the tab to be displayed
   final int currentIndex;
+
+  /// Whether another tab was requested to be displayed
   final ValueChanged<int>? onChanged;
 
+  /// The tabs to be displayed. This must have the same
+  /// length of [bodies]
   final List<Tab> tabs;
+
+  /// The bodies of the tabs. This must have the same
+  /// length of [tabs]
   final List<Widget> bodies;
 
-  final bool showNewButton;
+  /// Called when the new button is pressed or when the
+  /// shortcut `Ctrl + T` is executed.
+  ///
+  /// If null, the new button won't be displayed
   final void Function()? onNewPressed;
 
+  /// The icon of the new button
   final IconData addIconData;
+
+  /// Whenever the new button should be displayed.
+  bool get showNewButton => onNewPressed != null;
+
+  /// Whether the following shortcuts are enabled:
+  ///
+  /// - Ctrl + T to create a new tab
+  /// - Ctrl + F4 or Ctrl + W to close the current tab
+  /// - `1` to `8` to navigate through tabs
+  /// - `9` to navigate to the last tab
+  final bool shortcutsEnabled;
 
   @override
   void debugFillProperties(DiagnosticPropertiesBuilder properties) {
@@ -49,7 +86,7 @@ class TabView extends StatelessWidget {
   Widget build(BuildContext context) {
     debugCheckHasFluentTheme(context);
     final divider = SizedBox(
-      height: kTileHeight,
+      height: _kTileHeight,
       child: Divider(
         direction: Axis.vertical,
         style: DividerStyle(
@@ -57,12 +94,12 @@ class TabView extends StatelessWidget {
         ),
       ),
     );
-    return Acrylic(
+    Widget tabBar = Acrylic(
       child: Column(children: [
         Container(
           margin: EdgeInsets.only(top: 4.5),
           padding: EdgeInsets.only(left: 8),
-          height: kTileHeight,
+          height: _kTileHeight,
           width: double.infinity,
           child: Row(children: [
             ...tabs.map((e) {
@@ -115,26 +152,109 @@ class TabView extends StatelessWidget {
         if (bodies.isNotEmpty) Expanded(child: bodies[currentIndex]),
       ]),
     );
+    if (shortcutsEnabled)
+      return Shortcuts(
+        shortcuts: {
+          /// Ctrl + F4 or Ctrl + W closes the current tab
+          LogicalKeySet(
+            LogicalKeyboardKey.control,
+            LogicalKeyboardKey.f4,
+          ): _TabCloseIntent(),
+          LogicalKeySet(
+            LogicalKeyboardKey.control,
+            LogicalKeyboardKey.keyW,
+          ): _TabCloseIntent(),
+
+          /// Ctrl + T creates a new tab
+          LogicalKeySet(
+            LogicalKeyboardKey.control,
+            LogicalKeyboardKey.keyT,
+          ): _OpenNewTabIntent(),
+
+          /// Ctrl + (number from 1 to 8) navigate to that tab
+          /// Ctrl + 9 navigates to the last tab
+          /// TODO(bdlukaa): Ctrl + number. Currently blocked by https://github.com/bdlukaa/fluent_ui/issues/15
+          ...Map.fromIterable(
+            List.generate(9, (index) => index),
+            key: (number) {
+              final digits = [
+                LogicalKeyboardKey.digit1,
+                LogicalKeyboardKey.digit2,
+                LogicalKeyboardKey.digit3,
+                LogicalKeyboardKey.digit4,
+                LogicalKeyboardKey.digit5,
+                LogicalKeyboardKey.digit6,
+                LogicalKeyboardKey.digit7,
+                LogicalKeyboardKey.digit8,
+                LogicalKeyboardKey.digit9,
+              ];
+              return LogicalKeySet(digits[number]);
+            },
+            value: (number) => _ChangeTabIntent(number),
+          ),
+        },
+        child: Actions(
+          actions: {
+            _TabCloseIntent: CallbackAction<_TabCloseIntent>(onInvoke: (_) {
+              tabs[currentIndex].onClosed?.call();
+            }),
+            _OpenNewTabIntent: CallbackAction<_OpenNewTabIntent>(onInvoke: (_) {
+              onNewPressed?.call();
+            }),
+            _ChangeTabIntent: CallbackAction<_ChangeTabIntent>(onInvoke: (i) {
+              final tab = i.tab;
+              if (tab == 8) {
+                onChanged?.call(tabs.length - 1);
+              } else {
+                if (tabs.length - 1 >= tab) onChanged?.call(tab);
+              }
+            }),
+          },
+          child: Focus(child: tabBar),
+        ),
+      );
+    return tabBar;
   }
 }
 
-const double kTileWidth = 240.0;
-const double kTileHeight = 35.0;
+class _TabCloseIntent extends Intent {}
+
+class _OpenNewTabIntent extends Intent {}
+
+class _ChangeTabIntent extends Intent {
+  final int tab;
+
+  _ChangeTabIntent(this.tab);
+}
+
+const double _kTileWidth = 240.0;
+const double _kTileHeight = 35.0;
 
 class Tab {
   const Tab({
     Key? key,
     this.icon = const FlutterLogo(),
     required this.text,
-    this.closeIcon,
-    this.semanticsLabel,
+    this.closeIcon = Icons.close,
+    this.onClosed,
+    this.semanticLabel,
   });
 
+  /// The leading icon of the tab. [FlutterLogo] is used by default
   final Widget? icon;
-  final Widget text;
-  final Widget? closeIcon;
 
-  final String? semanticsLabel;
+  /// The text of the tab. Usually a [Text] widget
+  final Widget text;
+
+  /// The close icon of the tab. Usually an [IconButton] widget
+  final IconData? closeIcon;
+
+  /// Called when the close button is called or when the
+  /// shortcut `Ctrl + T` or `Ctrl + F4` is executed
+  final VoidCallback? onClosed;
+
+  /// {@macro fluent_ui.controls.inputs.HoverButton.semanticLabel}
+  final String? semanticLabel;
 }
 
 class _Tab extends StatelessWidget {
@@ -143,28 +263,29 @@ class _Tab extends StatelessWidget {
     Key? key,
     this.onPressed,
     required this.selected,
+    this.focusNode,
   }) : super(key: key);
 
   final Tab tab;
   final bool selected;
   final void Function()? onPressed;
-  final FocusNode focusNode = FocusNode();
+  final FocusNode? focusNode;
 
   @override
   Widget build(BuildContext context) {
     debugCheckHasFluentTheme(context);
     final style = context.theme;
     return HoverButton(
-      semanticsLabel: tab.semanticsLabel,
+      semanticLabel: tab.semanticLabel,
       focusNode: focusNode,
       cursor: selected ? (_) => SystemMouseCursors.basic : null,
       onPressed: onPressed,
       builder: (context, state) {
         Widget child = Container(
-          height: kTileHeight,
+          height: _kTileHeight,
           constraints: BoxConstraints(
-            maxWidth: kTileWidth,
-            minWidth: kTileWidth / 4,
+            maxWidth: _kTileWidth,
+            minWidth: _kTileWidth / 4,
           ),
           padding: EdgeInsets.symmetric(horizontal: 10, vertical: 6),
           decoration: BoxDecoration(
@@ -183,44 +304,50 @@ class _Tab extends StatelessWidget {
             if (tab.closeIcon != null)
               Theme(
                 data: style.copyWith(Style(
-                  iconButtonStyle: style.iconButtonStyle?.copyWith(
-                    IconButtonStyle(
-                      decoration: (state) {
-                        late Color? color;
-                        if (state.isNone)
-                          color = null;
-                        else
-                          color = ButtonStyle.buttonColor(style, state);
-                        return BoxDecoration(
-                          borderRadius: BorderRadius.circular(2),
-                          border: Border.all(style: BorderStyle.none),
-                          color: color,
-                        );
-                      },
-                      margin: EdgeInsets.zero,
-                      padding: EdgeInsets.zero,
-                      iconStyle: (state) => IconStyle(
-                        size: 20,
-                        color: () {
-                          if (state.isDisabled || state.isNone)
-                            return context.theme.disabledColor;
-                          else
-                            return context.theme.inactiveColor;
-                        }(),
-                      ),
+                  focusStyle: FocusStyle(
+                    primaryBorder: BorderSide(
+                      width: 1,
+                      color: style.inactiveColor ?? Colors.transparent,
                     ),
                   ),
                 )),
-                child: tab.closeIcon!,
+                child: IconButton(
+                  icon: Icon(tab.closeIcon),
+                  onPressed: tab.onClosed,
+                  style: IconButtonStyle(
+                    decoration: (state) {
+                      late Color? color;
+                      if (state.isNone)
+                        color = null;
+                      else
+                        color = ButtonStyle.buttonColor(style, state);
+                      return BoxDecoration(
+                        borderRadius: BorderRadius.circular(2),
+                        border: Border.all(style: BorderStyle.none),
+                        color: color,
+                      );
+                    },
+                    margin: EdgeInsets.zero,
+                    padding: EdgeInsets.zero,
+                    iconStyle: (state) => IconStyle(
+                      size: 20,
+                      color: () {
+                        if (state.isDisabled || state.isNone)
+                          return context.theme.disabledColor;
+                        else
+                          return context.theme.inactiveColor;
+                      }(),
+                    ),
+                  ),
+                ),
               ),
           ]),
         );
         return Semantics(
           selected: selected,
-          child: FocusBorder(
-            child: child,
-            focused: state.isFocused,
-          ),
+          focusable: true,
+          focused: state.isFocused,
+          child: FocusBorder(child: child, focused: state.isFocused),
         );
       },
     );